Create .kt
Explanation:

A class called "Person" is defined with two properties (name and age) and a method called "printDetails".
An object of the Person class is created using the "val" keyword.
The properties of the person1 object are set using the object reference and the dot (.) operator.
The printDetails() method of the person1 object is called using the object reference and the dot (.) operator. This method prints the person's name and age to the console.